// autogpt_platform/frontend/src/app/api/__generated__/models/graphExecutionsPaginated.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { GraphExecutionMeta } from "./graphExecutionMeta";
import type { Pagination } from "./pagination";

/**
 * Response schema for paginated graph executions.
 */
export interface GraphExecutionsPaginated {
  executions: GraphExecutionMeta[];
  pagination: Pagination;
}